cleanup, moved stages into their own directory, addressed most of the flake8 warnings, added better instructions for setting up Dahlia
