+++ fud/fud/utils.py
@@ -1,4 +1,5 @@
 import sys
 
+
 def eprint(*args, **kwargs):
     print(*args, **kwargs, file=sys.stderr)
